Add Google sign in token checking server side
Add verify_user function to verify that the google user is who they say
they are.

Also improve logged_in function, so it now returns the user DB object if
the user is logged in, or None if they aren't.

Add a userID field to the User model, to store the unique key for every
google account. This is better than using emails because emails can
change.